connect: Add CAManager for synchronizing CA operations
