+++ agent/consul/connect_ca_endpoint.go
@@ -165,18 +165,13 @@ func (s *ConnectCA) ConfigurationSet(
 
 	// If this is a secondary, check if the intermediate needs to be regenerated.
 	if s.srv.config.Datacenter != s.srv.config.PrimaryDatacenter {
-		// Get the current root certs from the primary DC.
-		var roots structs.IndexedCARoots
-		rootArgs := structs.DCSpecificRequest{
-			Datacenter: s.srv.config.PrimaryDatacenter,
+		// Attempt to take the CA lock in order to update the config.
+		if err := s.srv.caManager.setState(CAStateReconfig); err != nil {
+			return err
 		}
-		if err := s.srv.forwardDC("ConnectCA.Roots", s.srv.config.PrimaryDatacenter, &rootArgs, &roots); err != nil {
-			return fmt.Errorf("Error retrieving the primary datacenter's roots: %v", err)
-		}
+		defer s.srv.caManager.setReady()
 
-		s.srv.caProviderReconfigurationLock.Lock()
-		defer s.srv.caProviderReconfigurationLock.Unlock()
-		if err := s.srv.initializeSecondaryCA(newProvider, roots, args.Config); err != nil {
+		if err := s.srv.caManager.initializeSecondaryCA(newProvider, args.Config); err != nil {
 			return fmt.Errorf("Error updating secondary datacenter CA config: %v", err)
 		}
 		cleanupNewProvider = false
@@ -226,7 +221,7 @@ func (s *ConnectCA) ConfigurationSet(
 
 		// If the config has been committed, update the local provider instance
 		cleanupNewProvider = false
-		s.srv.setCAProvider(newProvider, newActiveRoot)
+		s.srv.caManager.setCAProvider(newProvider, newActiveRoot)
 
 		s.logger.Info("CA provider config updated")
 
@@ -239,7 +234,7 @@ func (s *ConnectCA) ConfigurationSet(
 
 	// First up, sanity check that the current provider actually supports
 	// cross-signing.
-	oldProvider, _ := s.srv.getCAProvider()
+	oldProvider, _ := s.srv.caManager.getCAProvider()
 	if oldProvider == nil {
 		return fmt.Errorf("internal error: CA provider is nil")
 	}
@@ -323,7 +318,7 @@ func (s *ConnectCA) ConfigurationSet(
 	// If the config has been committed, update the local provider instance
 	// and call teardown on the old provider
 	cleanupNewProvider = false
-	s.srv.setCAProvider(newProvider, newActiveRoot)
+	s.srv.caManager.setCAProvider(newProvider, newActiveRoot)
 
 	if err := oldProvider.Cleanup(); err != nil {
 		s.logger.Warn("failed to clean up old provider", "provider", config.Provider)
@@ -457,7 +452,7 @@ func (s *ConnectCA) SignIntermediate(
 		return acl.ErrPermissionDenied
 	}
 
-	provider, _ := s.srv.getCAProvider()
+	provider, _ := s.srv.caManager.getCAProvider()
 	if provider == nil {
 		return fmt.Errorf("internal error: CA provider is nil")
 	}
